Remove AclLinesAnswerElementInterface
